css/css-text-decor/text-decoration-serialization.html

<!DOCTYPE HTML>
<title>text-decoration shorthand serialization</title>
<link rel="help" href="https://drafts.csswg.org/css-text-decor-3/#text-decoration-property">
<link rel="help" href="https://drafts.csswg.org/cssom/#serialize-a-css-declaration-block">
<script type="text/javascript" src="/resources/testharness.js"></script>
<script type="text/javascript" src="/resources/testharnessreport.js"></script>
<div style="text-decoration: underline"></div>
<script>
test(() => {
  const style = getComputedStyle(document.querySelector('div'));
  assert_equals(style.getPropertyValue("text-decoration"), "underline solid rgb(0, 0, 0)");
});
</script>
